Implementacion funcion dados
